requirements.txt
altgraph==0.17
appdirs==1.4.3
APScheduler==3.6.3
asgiref==3.2.7
attrs==19.3.0
Automat==20.2.0
bcrypt==3.1.7
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
constantly==15.1.0
cryptography==2.9
cssselect==1.1.0
defusedxml==0.6.0
distlib==0.3.0
Django==3.0.5
django-cors-headers==3.2.1
emailsend==1.1
filelock==3.0.12
Flask==1.1.2
Flask-Compress==1.4.0
Flask-SQLAlchemy==2.4.1
future==0.18.2
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.11.2
logparser==0.8.2
lxml==4.5.0
MarkupSafe==1.1.1
MyApplication==0.1.0
mysql==0.0.2
mysqlclient==1.4.6
parsel==1.5.2
pbr==5.4.5
pefile==2019.4.18
pexpect==4.8.0
Pillow==7.1.1
priority==1.3.0
Protego==0.1.16
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
PyInstaller==3.6
pymongo==3.10.1
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pypiwin32==223
pyserial==3.4
pytesseract==0.3.4
pytz==2019.3
pywin32==227
pywin32-ctypes==0.2.0
queuelib==1.5.0
requests==2.23.0
Scrapy==2.0.1
scrapyd==1.2.1
scrapydweb==1.4.0
selenium==3.141.0
service-identity==18.1.0
six==1.14.0
SOAPpy==0.12.22
SQLAlchemy==1.3.16
sqlparse==0.3.1
Twisted==20.3.0
tzlocal==2.0.0
urllib3==1.25.8
w3lib==1.21.0
Werkzeug==1.0.1
wstools==0.4.8
zope.interface==5.1.0
